backend/requirements.txt
fastapi
uvicorn[standard]
google-genai>=1.5.0
python-dotenv
websockets
